requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
fastembed==0.6.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
huggingface-hub==0.29.3
humanfriendly==10.0
idna==3.10
kiwisolver==1.4.8
loguru==0.7.3
matplotlib==3.10.1
mmh3==5.1.0
mpmath==1.3.0
numpy==2.2.4
onnxruntime==1.21.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==6.30.1
py_rust_stemmers==0.1.5
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
seaborn==0.13.2
six==1.17.0
sympy==1.13.3
tokenizers==0.21.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
